Updated user interface
Length of text field now reflects the length of the longest word in the reference file. The text field now processes the user input when the user presses the "Enter" key. Fixed horizontal and vertical alignment of possible words. Made UserInterface into more of a class instead of just running it as the main method.